% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMSCRC.R
\name{classifyMSCRC}
\alias{classifyMSCRC}
\title{classify multi-omics subtypes for colorectal cancer (MSCRC).
This function classifies MSCRC for colorectal cancer, based on any combination of mRNA expression, microRNA expression, and DNA methylation data.}
\usage{
classifyMSCRC(exp)
}
\arguments{
\item{exp}{a data frame with samples in rows and multi-omics features in columns. The column names of exp should be X1, X2, ..., X196.}
}
\value{
A list label: a vector indicating the predicted subtype for each sample. prob: A data frame where each row represents a sample and each column represents the predicted probability for each subtype. multi-omics subtypes for colorectal cancer (MSCRC)
}
\description{
classify multi-omics subtypes for colorectal cancer (MSCRC).
This function classifies MSCRC for colorectal cancer, based on any combination of mRNA expression, microRNA expression, and DNA methylation data.
}
\examples{
result = classifyMSCRC(example_exp)
label = result$label
prob = result$probs


}
