% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_exp.R
\docType{data}
\name{example_exp}
\alias{example_exp}
\title{Example Expression Data}
\format{
A data frame with samples in rows and multi-omics features in columns.
}
\source{
the scaled expression of mRNA data (single omics data).
}
\usage{
example_exp
}
\description{
A dataset containing example expression data for testing.
}
\keyword{datasets}
