% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE39582_expr}
\alias{GSE39582_expr}
\title{Example Expression Data}
\format{
A data frame with z-score scaled values, samples in columns, genes in rows, rownames corresponding to gene symbols.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE39582}
}
\usage{
GSE39582_expr
}
\description{
A dataset containing example expression data for testing.
}
\keyword{datasets}
