% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{load model
This function load model for subtype prediction.}
\usage{
load_model()
}
\value{
A model.
}
\description{
load model
This function load model for subtype prediction.
}
