% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methyscaled}
\alias{methyscaled}
\title{DNA methylation profile of TCGA-COAD and TCGA-READ.}
\format{
A DNA methylation profile with 358 gene symbols and 315 samples.
}
\usage{
methyscaled
}
\description{
A dataset containing the pre-processed DNA methylation profile of 315 TCGA colorectal samples.
}
\keyword{datasets}
