% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNAscaled}
\alias{miRNAscaled}
\title{microRNA exoression profile of TCGA-COAD and TCGA-READ.}
\format{
A microRNA exoression profile with 196 microRNAs and 315 samples.
}
\usage{
miRNAscaled
}
\description{
A dataset containing the pre-processed microRNA exoression profile of 315 TCGA colorectal samples.
}
\keyword{datasets}
