% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNAscaled}
\alias{mRNAscaled}
\title{Gene expression profile of TCGA-COAD and TCGA-READ.}
\format{
A microRNA exoression profile with 951 mRNAs and 315 samples.
}
\usage{
mRNAscaled
}
\description{
A dataset containing the pre-processed mRNA exoression profile of 315 TCGA colorectal samples.
}
\keyword{datasets}
