% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projection_mxs}
\alias{projection_mxs}
\title{Projection matrixs derived from 315 piared mRNA expression, microRNA expression, and DNA methylation data from TCGA-COAD and TCGA-READ samples.}
\format{
A list derived from sparse mCCA using PMA::MultiCCA()
}
\usage{
projection_mxs
}
\description{
A list containing projection matrixs of each omics. In the list, under ws, the three elements correspond to the projection matrices for mRNA, miRNA, and methylation.
}
\keyword{datasets}
